requirements.txt
bcrypt
pycryptodome
pyperclip
PyQt5
